refactor: component separation
